clean clac_*.c
