+++ includes/calc.h
@@ -1,11 +1,8 @@
 #ifndef CALC_H
 # define CALC_H
 
-#include "utils.h"
 #include <math.h>
 #include "ogv1.h"
-
-#define	BOUND_ADJUSTMENT	0.001
 
 enum	e_axis
 {
@@ -13,41 +10,29 @@ enum	e_axis
 	HORIZONTAL,
 };
 
-typedef struct s_intersection
-{
-	enum e_axis	touching_axis; //接している軸 vertical or horizontal
-	int			x;
-	int			y;
-	double		distance;
-	double		wall_height;
-	int			origin_offset;
-	double		degree;
-} t_intersection;
-
 typedef struct s_inter
 {
 	t_pos		pos;
 	double		distance;
-	enum e_axis	axis; //接している軸 vertical or horizontal
+	enum e_axis	axis;
 	double		angle;
-	int			origin_offset; //or orign_x & origin_y
+	int			origin_offset;
 	double		wall_height;
 } t_inter;
 
-// calc_inter.c
-t_inter new_calc_inter(t_mlx *mlx, double ray_angle);
+//calc_interseciton.c
+t_inter calc_intersection(t_mlx *mlx, double ray_angle);
 
-// calc_inter_utils.c
-void draw_rect_safely(t_mlx *mlx, t_pos pos, size_t rect_size, int color);
-// double calc_wall_height(t_inter inter, t_player player);
+//calc_utils.c
+double sin_wrap(double angle);
+double cos_wrap(double angle);
+double tan_wrap(double angle);
+double cot_wrap(double angle);
 
-// calc_compare_make_inter.c
-t_inter	compare_make_inter(t_mlx *mlx, t_pos *v_inter_pos, t_pos *h_inter_pos, double ray_angle);
-
-// calc_search_horizontal_inter.c
-t_pos *search_horizontal_inter(t_mlx *mlx, double ray_angle);
-
-// calc_search_vertical_inter.c
-t_pos *search_vertical_inter(t_mlx *mlx, double ray_angle);
+//calc_utils2.c
+double fix_angle(double angle);
+t_pos	fix_move(t_mlx *mlx, t_pos pos, double move_angle, int step);
+bool	is_up(double degree);
+bool	is_right(double degree);
 
 #endif
